Permit compiling examples for py2.7 with cargo
Ideally the testing scripts could build for both Python 3.6 and 2.7, so
I've adjusted the example Cargo.tomls to suit this.

This goes against the design for the examples to be copyable into new
projects, but `path = "../.."` already ruined that. I'm making the
comment that you need to change Cargo.toml far more obvious and include
more instructions.

In terms of a proper template users could start with, Cargo templates
would be nice, but we're not there yet...
https://internals.rust-lang.org/t/pre-rfc-cargo-templates/5056