Remove incorrect docs about null log_stream_name
The AWS Lambda docs state that context.log_stream_name is null if the
Lambda function doesn't have write permission to CloudWatch Logs. This
is concerning since the return type of LambdaContext::log_stream_name is
&str.

However this appears to not actually be the case (for now, at least).
When removing CloudWatch Logs permissions from a Lambda execution role,
a log stream name was still generated (although not written to).